+++ modules/polyfills/test/images-node/images-node.spec.js
@@ -1,18 +1,14 @@
 /* eslint-disable max-len */
 import test from 'tape-promise/tape';
 import '@loaders.gl/polyfills';
+import {isBrowser} from '@loaders.gl/polyfills/utils/globals';
 
 /* global _encodeImageNode, _parseImageNode */
-test('TextEncoder', t => {
-  t.equals(
-    typeof _encodeImageNode,
-    'function',
-    '_encodeImageNode successfully installed on global'
-  );
-  t.end();
-});
+test('Node image polyfills', t => {
+  if (!isBrowser) {
+    t.equals(typeof _encodeImageNode, 'function', 'global._encodeImageNode successfully installed');
 
-test('TextDecoder', t => {
-  t.equals(typeof _parseImageNode, 'function', '_parseImageNode successfully installed on global');
+    t.equals(typeof _parseImageNode, 'function', 'global._parseImageNode successfully installed');
+  }
   t.end();
 });
